lateral-synthesis/requirements.txt
mcp>=1.6.0
wonderwords>=2.2.0
pyyaml>=6.0
